+++ src/tests.rs
@@ -192,94 +192,6 @@ fn test_parse_numeric_input_empty() {
 }
 
 #[test]
-fn test_filter_by_max_length() {
-    use super::*;
-    use crate::filtering::ListFilter;
-
-    let constraints = ConstraintList::_new(Rc::new(RefCell::new(vec![
-        vec![0; 10],
-        vec![0; 3],
-        vec![0; 5],
-    ])));
-    let mut filter: ListFilter = ListFilter::new(constraints.clone());
-
-    filter.by_max_length(4);
-    let filtered = filter.get_filtered(0, 50).0;// filter.get_filtered(0, 50);
-    assert_eq!(filtered.len(), 1);
-
-    filter.by_max_length(5);
-    let filtered2 = filter.get_filtered(0, 50).0;
-    assert_eq!(filtered2.len(), 2);
-
-    filter.by_max_length(1);
-    let filtered3 = filter.get_filtered(0, 50).0;
-    assert_eq!(filtered3.len(), 0)
-}
-
-#[test]
-fn test_filter_by_cell() {
-    use super::*;
-    use crate::filtering::ListFilter;
-
-    let constraints = ConstraintList::_new(Rc::new(RefCell::new(vec![
-        vec![1; 10],
-        vec![10; 3],
-        vec![10; 3],
-    ])));
-    let mut filter: ListFilter = ListFilter::new(constraints.clone());
-    filter.reinit();
-
-    filter.by_cell(1, 1);
-    let filtered = filter.get_filtered(0, 50).0;
-    assert_eq!(filtered.len(), 1);
-
-    filter.by_cell(1, 2);
-    let filtered2 = filter.get_filtered(0, 50).0;
-    assert_eq!(filtered2.len(), 2);
-
-    filter.by_cell(2, 2);
-    let filtered3 = filter.get_filtered(0, 50).0;
-    assert_eq!(filtered3.len(), 0);
-}
-
-#[test]
-fn test_clear_filters_and_multiple_filters() {
-    use super::*;
-    use crate::filtering::ListFilter;
-
-    let constraints = ConstraintList::_new(Rc::new(RefCell::new(vec![
-        vec![1; 10],
-        vec![1; 3],
-        vec![10; 3],
-    ])));
-    let mut filter: ListFilter = ListFilter::new(constraints);
-    filter.reinit();
-
-    filter.by_cell(1, 1);
-    let filtered = filter.get_filtered(0, 50).0;
-    assert_eq!(filtered.len(), 2);
-    filter.by_max_length(3);
-    let filtered2 = filter.get_filtered(0, 50).0;
-    assert_eq!(filtered2.len(), 1);
-    filter.clear_cell();
-    let cleared = filter.get_filtered(0, 50).0;
-    assert_eq!(cleared.len(), 2);
-    filter.clear_length();
-    let cleared2 = filter.get_filtered(0, 50).0;
-    assert_eq!(cleared2.len(), 3);
-
-    let _ = filter.by_cell(1, 1);
-    filter.by_max_length(3);
-    let filtered4 = filter.get_filtered(0, 50).0;
-    assert_eq!(filtered4.len(), 1);
-    // filter.clear_all();
-    filter.clear_length();
-    filter.clear_cell();
-    let cleared3 = filter.get_filtered(0, 50).0;
-    assert_eq!(cleared3.len(), 3);
-}
-
-#[test]
 fn test_constraint_list() {
     use super::*;
